% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_code.R
\name{benchmark}
\alias{benchmark}
\title{Benchmark}
\usage{
benchmark(x)
}
\arguments{
\item{x}{Numeric or integer to calculate the factorial from}
}
\value{
Returns a table with the time it cost to produce each result
}
\description{
Comparison made between the four different implementations with the microbenchmark package.
}
\examples{
Factorial_loop(5)

}
