% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_code.R
\name{Factorial_func}
\alias{Factorial_func}
\title{Factorial Recursion}
\usage{
Factorial_func(x)
}
\arguments{
\item{x}{Numeric or integer to calculate the factorial from}
}
\value{
Returns the factorial
}
\description{
This is a simple function that calculates the Factorial using recursion
}
\examples{
Factorial_func(5)

}
