% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_code.R
\name{Factorial_loop}
\alias{Factorial_loop}
\title{Factorial Loop}
\usage{
Factorial_loop(x)
}
\arguments{
\item{x}{Numeric or integer to calculate the factorial from}
}
\value{
Returns the factorial
}
\description{
This is a simple function that calculates the Factorial using a loop
}
\examples{
Factorial_loop(5)

}
