% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_code.R
\name{Factorial_mem}
\alias{Factorial_mem}
\title{Factorial Memoization}
\usage{
Factorial_mem(x)
}
\arguments{
\item{x}{Numeric or integer to calculate the factorial from}
}
\value{
Returns the factorial
}
\description{
This is a simple function that calculates the Factorial using memoization
}
\details{
In order for this function to work you have to create a table previously
fac_tbl <- c(rep(NA,x))
}
\examples{
Factorial_mem(5)

}
