% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_code.R
\name{Factorial_reduce}
\alias{Factorial_reduce}
\title{Factorial Reduce}
\usage{
Factorial_reduce(x)
}
\arguments{
\item{x}{Numeric or integer to calculate the factorial from}
}
\value{
Returns the factorial
}
\description{
This is a simple function that calculates the Factorial using reduce
}
\examples{
Factorial_reduce(5)

}
